NLP-Backend/GDELT/requirements.txt
requests==2.31.0
pandas>=2.2.0
pandas-gbq>=0.22.0
google-cloud-bigquery>=3.0.0
newspaper3k>=0.2.8
nltk>=3.8.1
lxml_html_clean